requirements.txt
discord.py==1.7.3
discord-py-interactions==4.4.1
discord-py-slash-command==1.1.1
python-dotenv~=1.0.1
pytz~=2024.1
requests~=2.31.0